Maximum Float number

+++ FindMax.cs
@@ -8,7 +8,7 @@ namespace Generics
 {
     internal class FindMax
     {
-        public static void FindMaxValue(int first, int second, int third)
+        public static void FindMaxValue(float first, float second, float third)
         {
             if (first.CompareTo(second) >= 0 && first.CompareTo(third) >= 0)
             {
